+++ rubyllm_chat_app/app/controllers/chat_cards_controller.rb
@@ -1,5 +1,9 @@
 class ChatCardsController < ApplicationController
   def index
-    @chats = Chat.all.order(updated_at: :desc)
+    if params[:show_all]
+      @chats = current_user.chats.order(updated_at: :desc)
+    else
+      @chats = current_user.chats.with_attached_generated_image.order(updated_at: :desc)
+    end
   end
 end