+++ rubyllm_chat_app/app/models/chat.rb
@@ -10,6 +10,8 @@ class Chat < ApplicationRecord
   has_one_attached :generated_image # , prefix: 'chats/generated_images' # not implemented see https://github.com/rails/rails/issues/32790
   has_many :messages, -> { order(created_at: :asc) }, dependent: :destroy # Ensure messages are ordered
 
+  scope :with_attached_generated_image, -> { joins(:generated_image_attachment) }
+
   # Riccaredo this WONT work as its in PostgreS vincolo.
   #belongs_to :user, optional: true # Example